Build package on RISCV64, drop atlas build-dep, fix twice-build
RISCV build really didn't work with LLVM16,
but hopefully things are better in LLVM17.
It seems both lld and libomp packages exist for RISCV,
so let's build with those enabled.

As for atlas, it was a build dep only for two arches,
for which the package isn't even allowed to be built,
so we can safely drop it.

As for build-twice, indeed, our build tree was never removed
after build, so doing that (and moving it out of the source dir!)
is helpful.